Added all the files

+++ Cogs/Tools.py
@@ -10,14 +10,19 @@ class Tools(commands.Cog):
     def __init__(self, bot):
         self.bot = bot
 
+    @nextcord.slash_command(name='dm', description='DMs a user', guild_ids=[927811764895227945, 994484729917210646])
+    async def dm(self, interaction: nextcord.Interaction, member: nextcord.Member, message: str):
+        await member.send(f'**Message from {interaction.user.mention}:**\n{message}')
+        await interaction.response.send_message(f'Sent message to {member.mention}', ephemeral = True)
+    
     @commands.command(name='managenickname', aliases = ['mn'])
-    @commands.has_any_role('Admin', 'Head Admin', 'Manager')
+    @commands.has_any_role('Admin', 'Operations Manager', 'Team Leader', 'Deputy Team Leader', 'COO', 'COS', 'COM', 'SS')
     async def managenickname(self, ctx, member: nextcord.Member, *, nick: str):
         await member.edit(nick=nick)
         await ctx.send(f"{member.mention}'s nickname has been changed to {nick}!")
     
     @commands.command(name='poll')
-    @commands.has_any_role('Events Team', 'Support Team', 'Development Team')
+    @commands.has_any_role('Events Team', 'Support Team', 'Development Team', 'COO', 'COS', 'COM', 'SS')
     async def poll(self, ctx, question: str, option1: str = None, option2: str = None):
         embed = nextcord.Embed(title=f'**{question}**',
                                description=f'<:one:988434351614218270> {option1}\n<:two:988434728090734623> {option2}')
@@ -29,9 +34,9 @@ class Tools(commands.Cog):
         await msg.add_reaction(e2)
 
     @commands.command(name='addcrew', aliases=['ac'])
-    @commands.has_any_role('Admin', 'Head Admin', 'Manager')
+    @commands.has_any_role('Operations Manager', 'Team Leader', 'COO', 'COS', 'COM', 'SS')
     async def addcrew(self, ctx, member: nextcord.Member):
-        role = nextcord.utils.get(member.guild.roles, name='Helper')
+        role = nextcord.utils.get(member.guild.roles, name='Server Staff')
         await member.add_roles(role)
         embed = nextcord.Embed(title=f'Hired {member}!', description=f'{member.mention} Welcome to the staff team!')
         await ctx.send(embed=embed)
@@ -46,7 +51,7 @@ class Tools(commands.Cog):
             await ctx.reply(embed=embed)
 
     @commands.command(name='remcrew', aliases=['rc'])
-    @commands.has_any_role('Admin', 'Head Admin', 'Manager')
+    @commands.has_any_role('Operations Manager', 'Team Leader', 'COO', 'COS', 'COM', 'SS')
     async def remcrew(self, ctx, member: nextcord.Member):
         role = nextcord.utils.get(member.guild.roles, name='Helper')
         await member.remove_roles(role)
@@ -65,7 +70,7 @@ class Tools(commands.Cog):
             await ctx.reply(embed=embed)
 
     @commands.command(name='lock')
-    @commands.has_any_role('Admin', 'Head Admin', 'Manager')
+    @commands.has_any_role('Sr. Moderator', 'Admin', 'Operations Manager', 'COO', 'COS', 'COM', 'SS')
     async def lock(self, ctx, channel: nextcord.TextChannel = None):
         channel = ctx.channel or channel
         overwrite = channel.overwrites_for(ctx.guild.default_role)
@@ -82,7 +87,7 @@ class Tools(commands.Cog):
             return
 
     @commands.command(name='unlock')
-    @commands.has_any_role('Admin', 'Head Admin', 'Manager')
+    @commands.has_any_role('Sr. Moderator', 'Admin', 'Operations Manager', 'COO', 'COS', 'COM', 'SS')
     async def unlock(self, ctx, channel: nextcord.TextChannel = None):
         channel = channel or ctx.channel
         overwrite = channel.overwrites_for(ctx.guild.default_role)
@@ -99,7 +104,7 @@ class Tools(commands.Cog):
             return
 
     @commands.command(name='slowmode', aliases=['slow', 'sm'])
-    @commands.has_any_role('Moderation Team', 'Manager')
+    @commands.has_any_role('Sr. Moderator', 'Operations Manager', 'COO', 'COS', 'COM', 'SS')
     async def slowmode(self, ctx, time: int):
         await ctx.channel.edit(slowmode_delay=time)
         await ctx.send(f'Slowmode set to {time} seconds!')
@@ -125,7 +130,7 @@ class Tools(commands.Cog):
         await ctx.send(f'The members for this role are: \n{role_members}')
     
     @commands.command(name='role')
-    @commands.has_any_role('Admin', 'Head Admin', 'Manager')
+    @commands.has_any_role('Operations Manager', 'Team Leader', 'COO', 'COS', 'COM', 'SS')
     async def role(self, ctx, member: nextcord.Member, role_name: str):
         role = utils.get(member.guild.roles, name=role_name)
         author = ctx.author
@@ -145,7 +150,7 @@ class Tools(commands.Cog):
             await ctx.send("Hm..... The command doesn't seem to work <:thinking:>")
 
     @commands.command(name='removerole', aliases=['remrole'])
-    @commands.has_any_role('Admin', 'Head Admin', 'Manager')
+    @commands.has_any_role('Admin', 'Operations Manager', 'COO', 'COS', 'COM', 'SS')
     async def removerole(self, ctx, member: nextcord.Member, role_name: str):
         role = utils.get(member.guild.roles, name=role_name)
         author = ctx.author
@@ -166,7 +171,7 @@ class Tools(commands.Cog):
 
 
     @commands.command(name='nuke')
-    @commands.has_any_role('Head Admin', 'Manager')
+    @commands.has_any_role('SS', 'COO', 'COS')
     async def nuke(self, ctx):
         channel = await ctx.channel.clone(name=f'{ctx.channel.name}')
         await ctx.channel.delete()
@@ -180,7 +185,7 @@ class Tools(commands.Cog):
             await ctx.send(embed=embed)
             return
 
-    @commands.group(name='space', aliases=['spc'])
+    @commands.group(name='ticket')
     async def space(self, ctx):
         await ctx.send('This is the testspace command. Sub-Commands are: Create, Add, Close')
 
@@ -194,13 +199,13 @@ class Tools(commands.Cog):
         }
 
         channel = await guild.create_text_channel(name=f'{ctx.author.name}', overwrites=overwrites, reason=reason,
-                                                  topic=f'Bot testing space for {ctx.author.name}')
+                                                  topic=f'Support Ticket')
         await channel.send(f'{user.mention}')
         await asyncio.sleep(2)
         await channel.purge(limit=1)
         embed1 = nextcord.Embed(title='Reason:', description=f'{reason}')
         await channel.send(embed=embed1)
-        embed = nextcord.Embed(title='Channel Created!', description='Your channel has been created!')
+        embed = nextcord.Embed(title='Channel Created!', description=f'Your ticket has been created! {channel.mention}')
         await ctx.send(embed=embed)
         await asyncio.sleep(5)
         await ctx.channel.purge(limit=3)
@@ -212,8 +217,11 @@ class Tools(commands.Cog):
         overwrite.read_messages = True
         overwrite.send_messages = True
         overwrite.view_channel = True
-        await channel.set_permissions(member, overwrite=overwrite)
-        await ctx.send(f'{member.name} has been added to the space!')
+        if channel.topic == 'Support Ticket':
+            await channel.set_permissions(member, overwrite=overwrite)
+            await ctx.send(f'{member.name} has been added to the ticket!')
+        else:
+            await ctx.send('This is not a support ticket!')
 
     @space.command(name='remove')
     async def remove(self, ctx, member: nextcord.Member):
@@ -222,18 +230,24 @@ class Tools(commands.Cog):
         overwrite.read_messages = False
         overwrite.send_messages = False
         overwrite.view_channel = False
-        await channel.set_permissions(member, overwrite=overwrite)
-        await ctx.send(f'{member.name} has been removed from the space!')
-    
+        if channel.topic == 'Support Ticket':
+            await channel.set_permissions(member, overwrite=overwrite)
+            await ctx.send(f'{member.name} has been removed from the space!')
+        else:
+            await ctx.send('This is not a support ticket!')
+        
     @space.command(name='close')
     @commands.has_role('Support Team')
     async def close(self, ctx):
         channel = ctx.channel
-        await ctx.send('Closing space.....')
-        await channel.delete()
+        if channel.topic == 'Support Ticket':
+            await ctx.send('Closing space.....')
+            await channel.delete()
+        else:
+            await ctx.send('This is not a support ticket!')
 
     @commands.command(name='task')
-    @commands.has_any_role('Admin', 'Head Admin', 'Manager')
+    @commands.has_any_role('Operations Manager', 'Team Leader', 'Deputy Team Leader', 'COO', 'COS', 'COM', 'SS')
     async def tasks(self, ctx, member: nextcord.Member, *, task: str):
         await ctx.send(f'{member.mention} has been assigned a task: {task}!')
 
@@ -246,7 +260,7 @@ class Tools(commands.Cog):
             return
 
     @commands.command(name='clear', aliases=['c', 'purge'])
-    @commands.has_any_role('Moderation Team', 'Manager')
+    @commands.has_any_role('Moderation Team', 'Operations Manager', 'COO', 'COS', 'COM', 'SS')
     async def clear(self, ctx, amount=2):
         await ctx.channel.purge(limit=amount + 1)
